[TASK] Implemented basic file access functionality
The local filesystem driver is now able to do basic file operations,
like creating and reading from/writing to files. All operations are unit
tested on driver-level, the upper layers still have to be implemented
and tested.

This commit introduces the new concept of a FileHandle. This handle is -
for normal files - basically a wrapper for the resource pointer. It
might be of good use later on, when dealing with more complex ways of
accessing (remote) files. The handle is meant to be returned by the
driver when a file is opened. It is then handed to all functions that
deal with manipulating files inside the driver. The driver then gets all
information neccessary for its file operations from the FileHandle. The
logic to manage the FileHandle inside the File class still has to be
implemented.