[TASK] Add Repository and Factory classes
The Repository will be the central hub for getting files and folders from
the VFS. It is part of the high-level API of the VFS.

The Factory is responsible for doing object creation and initialization.

This commit also brings a new abstraction class for the root node of the
VFS tree.