[TASK] Make Node aware of its path inside the mount
Node is now also aware of its parent and has a getter for the name. The
name should be extracted from the properties on object creation and maybe
also injectable via a setter.

Also made the method t3lib_vfs_driver_Local::getAbsolutePath() public
and changed it to use the new method in t3lib_vfs_Node.

Additionally introduces a new test in _driver_Local.