[TASK] Folder: implement getSubfolder()
The unit tests for this method require a patched PHPUnit and a patched
TYPO3 core. This is because the static attributes of classes are not
strictly isolated in unit tests, so the prepared statement in
getSubfolder() is preserved between unit tests. This makes the second
test fail as the prepared statement we use there is not the one we
intended for this test. This bug can be circumvented by running the
tests in their own processes; but this in turn triggers problems with
the way PHPUnit is run inside TYPO3: the testing process is spawned
with a custom script that resets the environment to the state it had
inside the parent process.
Because of the way TYPO3 is initialized when running from the command
line, exactly this process is repeated, but it fails, as there is
insufficient input for that. To prevent this, we have to prevent
execution of the CLI command (registered for cli_dispatch.phpsh) and
also must not execute the PHP command runner. For this, PHPUnit has to
set $TYPO3_dontExecuteCli inside the child process.

See http://forge.typo3.org/issues/13903 for the EXT:phpunit issue.