[TASK] Folder: implement getSubfolders()
The current implementation doesn't support matching patterns, which
would be a nice addition.

Also added a test for getSubfolder() that checks if an exception is
thrown if no subfolders are found. We implicitly checked this by
expecting an exception in another test (because it eased testing; we
didn't have to mock everything that was neccessary for folder creation),
but this was rather bad style as these two things are quite orthogonal
concerns. The extra test produces a much better feeling :-)