[+API][!!!] add method Node::getPath()
This method returns the complete path of a node. There is a special
method in Folder that always returns the path with a trailing slash.

Warning: This changes the behaviour of getPathInMountpoint(): if the
requested path does not include the current node, a slash is appended to
signify that the last path part does not point to a leaf in the node
tree.