[~API][!!!] Repository can insert new records
Method updateNodeInDatabase() is renamed to persistNodeToDatabase(), and
it now checks if a node is new. If it is, a new database row is created
and the uid is injected into the node.

Currently only the tstamp is automatically updated when creating or
updating a node. More properties are yet to come.