[+API] Drivers support file hashing
File hashing will be neccessary for e.g. checking if a file is equal to
another one, or if it has been modified (as modification times are quite
unreliable in many scenarios).

In principle, there is support for different hashing algorithms. This
is to be able to easily switch to a different algorithm in case SHA-1
(the only currently implemented hash) gets further compromised.