[~API] Node: add hasProperty()
The new method makes it easier to check for existing properties;
additionally, all available properties now have to be registered in
subclasses. This is a bit weird as it adds another place to update when
changing the database schema, but it is the easiest solution for now...

Based on the available properties, the Repository will be able to
properly update the database.