[+FEATURE] Add indexer class
The indexer can currently index single files and folders. Methods for
mass indexing are missing, but should be added soon. When indexing a
mass of files, there should be some clever mechanism for checking all
subnodes for their status if the folder is already indexed, otherwise we
will potentially throw a lot of queries to the database just to index a
single huge folder.