[~FEATURE] Indexing path fails if no node in path is indexed
Having a path where no no node is indexed means that there is no
mountpoint available in that path, as we don't have an implicit
mountpoint at the root node (which might point to some local storage).
This means we can only proceed with indexing if there is any mountpoint
(= storage root) inside the path.