[+BUGFIX] Repository does not set uid correctly
The uid of inserted nodes is not correctly set outside the unit tests,
because the Repository uses the wrong semantics for the query methods.
The method t3lib_DB::exec_INSERTquery does not return the uid of the
newly inserted record - t3lib_DB::sql_insert_id() should be queried for
that instead.